+++ src/server/address.js
@@ -1,9 +1,10 @@
-var ifaces = require('os').networkInterfaces(),
-    address = Object.values(ifaces)
+var networkInterfaces = require('os').networkInterfaces
+
+module.exports = (proto, port)=>{
+    var address = Object.values(networkInterfaces())
         .reduce((a,b)=>a.concat(b), [])
         .filter(i=>i.family === 'IPv4')
         .map(i=>i.address + ':')
 
-module.exports = (proto, port)=>{
     return address.map(x=>proto + x + port)
 }